// sourcecode/BTrak.Packages/snova-audits-module/projects/project-components/src/lib/audits/dependencies/models/testcasedropdown.ts
export class TestCaseDropdownList {
    id: string;
    value: string;
    description: string;
    isArchived: boolean;
    sectionId: string;
    auditCategoryId: string;
    statusHexValue: string;
    statusShortName: string;
    profileImage: string;
    isHierarchical: boolean;
    timeStamp: any;
}